requirement.txt
openai
requests
ruff
mypy
pandas
python-dotenv